Heavily refactor the Addr class.
This removes the ability to implicitly convert from an Addr to an
integer type (you must explicitly convert now), and removes the
ambiguous overloaded operators, as well as some operators that I think
we can live without until there's a pressing reason to add them back.

+++ src/DataObject.cpp
@@ -10,7 +10,7 @@ uint64_t DataObject::getSize() const { return this->Size; }
 
 void DataObject::toProtobuf(MessageType* Message) const {
   nodeUUIDToBytes(this, *Message->mutable_uuid());
-  Message->set_address(this->Address);
+  Message->set_address(static_cast<uint64_t>(this->Address));
   Message->set_size(this->Size);
 }
